refactor(backend): extract metric crud into crud/metric.py
Move the metric CRUD out of the crud/__init__.py monolith into a dedicated
crud/metric.py, following crud/explorer.py and the same split rule in
apps/backend/AGENTS.md. Pure move: every function body is byte-identical to
the original.

The block is self-contained - it calls nothing else in crud/__init__.py,
nothing else there calls into it, and its _METRIC_RELATED_FIELDS constant is
used only inside the block. It also frees no module-level imports, so the
import header is left byte-identical to main.

The transitional re-export sits at the extraction site rather than in the
import block, deliberately: it keeps the header untouched so this does not
collide with other in-flight splits of the same file. Removed in the next
commit.
